AAIDC_Project_1/Examples/requirements.txt
python-dotenv
langchain_groq
langchain-google-genai
google-generativeai
ollama
langchain-ollama
langchain_community
chromadb
langchain 
pyyaml
